requirements.txt
streamlit>=1.30.0
torch>=1.12
torchvision>=0.13
fastai>=2.7.13
Pillow
google-generativeai